added chat functionality

+++ src/lib.rs
@@ -4,4 +4,4 @@ pub mod networking;
 pub mod game_controller;
 
 pub type Color = (u8, u8, u8);
-pub type Position = (u8, u8);
+pub type Position = (usize, usize);
